requirements.txt
certifi==2020.6.20
chardet==3.0.4
cloudinary==1.20.0
Django==2.0.6
django-filter==1.1.0
django-sendgrid-v5==0.8.1
djangorestframework==3.11.0
future==0.18.2
idna==2.10
mysqlclient==2.0.1
pyfcm==1.4.7
PyJWT==1.7.1
python-http-client==3.3.1
pytz==2020.1
requests==2.24.0
sendgrid==6.4.7
six==1.15.0
starkbank-ecdsa==1.1.0
urllib3==1.25.10
